moved nav components into components
